+++ src/ReviewAutomation/Application/UseCases/Reviews/ApproveReviewUseCase.cs
@@ -1,6 +1,4 @@
-using Athos.ReviewAutomation.Core.DTOs;
-using Athos.ReviewAutomation.Infrastructure.Services;
-using Athos.ReviewAutomation.Models;
+using Athos.ReviewAutomation.Application.DTOs.Reviews;
 
 namespace Athos.ReviewAutomation.Application.UseCases.Reviews
 {
