+++ src/ReviewAutomation/Application/UseCases/Reviews/IApproveReviewUseCase.cs
@@ -1,5 +1,4 @@
-using Athos.ReviewAutomation.Core.DTOs;
-using Athos.ReviewAutomation.Models;
+using Athos.ReviewAutomation.Application.DTOs.Reviews;
 
 namespace Athos.ReviewAutomation.Application.UseCases.Reviews
 {
